Fix files endpoint (#537)
In UAT the files endpoint failed validation because we triggered a call to BGS, but lacked the required information to make a request. It turns out the call to BGS isn't even necessary, so we updated the call to only call BGS when needed.

We also updated the returned id from the endpoint to reflect the internal id, since that is what is needed to request the pdfs from the documents endpoint.