Add authentication to API (#539)
We need to add authentication to our two APIs. Unfortunately, since they are accessed differently, we need two different authentication mechanisms.

The /files endpoint is hit from the Caseflow backend, so we will use a static API key saved as an environment variable to authenticate.

The /documents endpoint is hit from the Caseflow frontend, so we will use the user token (since the token is shared between Caseflow and eFolder Express) to authenticate. We will just check that the token has the Reader role.

Furthermore, since the veteran id is PII we move it from a get parameter to a value passed in the header. We also pass the station_id and css_id in the header.

Also starting to return the vbms_document_id since we'll need that on the Caseflow side.