Add sensitivity check (#644)
Now we have a BGSService service that calls either Fakes::BGSService or ExternalApi::BGSService depending on if we're in production or not. BGSService no longer takes a user object to initialize, it derives the current user from the request store. To use the service, a client can just create a new BGSService and memoize it for future calls.

We removed some code that would use Fakes::BGSService in production if a demo eFolder was being looked up. I don't think anyone looks up demo cases in production, and putting demo data in production seems dangerous.

Finally, we added a staging environment for eFolder. To access it all you need to do is run rails s -e staging. You'll also need to create the db RAILS_ENV=staging rake db:setup.