Verify veteran_info for "not found", do not rely only on BGS return_message (#1196)
e.g. https://sentry.ds.va.gov/department-of-veterans-affairs/efolder/issues/8396/

Fixes VBMS::FilenumberDoesNotExist -- we were trusting too much in the preflight BGS return message value.

connects department-of-veterans-affairs/caseflow#10732 and department-of-veterans-affairs/caseflow#6680

This PR also adds local support for testing. If the file number supplied starts with 404 then it can be tested for "validity" as well as "not found".