🔥 Remove `uglifier`
During assets precompile in a 'production' environment, we encountered the following error:

  Uglifier::Error: Unexpected token: name (compare1).
  To use ES6 syntax, harmony mode must be enabled with Uglifier.new(:harmony => true).

Per the `uglifier` README:

  UglifyJS only works with ES5. If you need to compress ES6, `ruby-terser` is a better option.

Looking at the Caseflow git history for comparison, it looks like the `uglifier` gem was
removed in favor of using Webpack to perform JS compression via the `UglifyjsWebpackPlugin`.
Later, the `UglifyjsWebpackPlugin` was removed when Webpack v4 incorporated the
`TerserWebpackPlugin` out-of-the-box:
https://github.com/department-of-veterans-affairs/caseflow-efolder/blob/9853eaeb98692099f1e62435de9a4dc08292fa53/client/yarn.lock#L6119

It appears that there may need to be some additional configuration added to the
`webpack.config.js` file in order to leverage the Terser plugin:
https://v4.webpack.js.org/plugins/terser-webpack-plugin/

However, the Caseflow `webpack.config.js` does not include the Terser configuration at this time,
and so, in keeping parity with Caseflow, we will omit this configuration in eFolder as well
and leave it as a future exercise should it be necessary to enact JS compression.